Add model loader, initial work on data transfer

+++ src/rn/vki/trace.cpp
@@ -8,7 +8,7 @@ namespace rn::vki {
 
 std::map<std::thread::id, LastCall> stores{};
 
-uint32_t baseOffset = std::string_view{__FILE__}.length() - std::string_view{"src/rn/vki/trace.cpp"}.length();
+size_t baseOffset = std::string_view{__FILE__}.length() - std::string_view{"src/rn/vki/trace.cpp"}.length();
 
 void trace(uint32_t line, const std::string_view &file, const std::string_view &code) {
 	LastCall store;
